Refactor Native/Web specific code
