+++ Tools/WinML-Dashbaord/src/native/util.ts
@@ -1,0 +1,5 @@
+import * as fs from 'fs';
+
+export function isWeb() {
+    return !fs.exists;
+}
