+++ Tools/WinML-Dashbaord/src/view/edit/View.tsx
@@ -3,6 +3,7 @@ import * as React from 'react';
 
 import Resizable from '../../components/Resizable';
 import { ModelProtoSingleton } from '../../datastore/proto/modelProto';
+import { showWebOpenDialog } from '../../native/dialog';
 import LeftPanel from './LeftPanel';
 import * as netron from './netron/Netron';
 import RightPanel from './RightPanel';
@@ -14,8 +15,6 @@ interface IComponentState {
 }
 
 export default class EditView extends React.Component<{}, IComponentState> {
-    private openFileInput: React.RefObject<HTMLInputElement> = React.createRef();
-
     constructor(props: {}) {
         super(props);
         this.state = {
@@ -33,7 +32,6 @@ export default class EditView extends React.Component<{}, IComponentState> {
                 <Resizable>
                     <DefaultButton text='Open file' onClick={this.openFile}/>
                     <DefaultButton text='Save file' onClick={ModelProtoSingleton.download}/>
-                    <input type='file' style={{display: 'none'}} accept=".onnx,.pb,.meta,.tflite,.keras,.h5,.json,.mlmodel,.caffemodel" ref={this.openFileInput} />
                     <RightPanel />
                 </Resizable>
             </div>
@@ -41,24 +39,11 @@ export default class EditView extends React.Component<{}, IComponentState> {
     }
 
     private openFile = () => {
-        if (!this.openFileInput.current) {
-            return;
-        }
-        this.openFileInput.current.addEventListener('change', this.onFileSelected);
-        this.openFileInput.current.click();
-    }
-
-    private onFileSelected = () => {
-        const openFileInput = this.openFileInput.current;
-        if (!openFileInput) {
-            return;
-        }
-        const files = openFileInput.files;
-        if (!files || !files.length) {
-            return;
-        }
-        this.setState({
-            file: files[0],
-        })
+        showWebOpenDialog('.onnx,.pb,.meta,.tflite,.keras,.h5,.json,.mlmodel,.caffemodel')
+            .then((files) => {
+                if (files) {
+                    this.setState({ file: files[0] });
+                }
+            });
     }
 }
